Move Participant class to game package
