+++ ladder/jaeeun-na-1/src/main/java/org/ooptraining/render/policy/OneLineRenderPolicy.java
@@ -1,6 +1,6 @@
 package org.ooptraining.render.policy;
 
-import org.ooptraining.Participant;
+import org.ooptraining.game.Participant;
 import org.ooptraining.render.RenderContext;
 import org.ooptraining.render.RenderPolicy;
 import org.ooptraining.game.GameContext;
