+++ ladder/jaeeun-na-1/src/main/java/org/ooptraining/util/StringUtils.java
@@ -1,6 +1,6 @@
 package org.ooptraining.util;
 
-import org.ooptraining.Participant;
+import org.ooptraining.game.Participant;
 import org.ooptraining.game.GameContext;
 
 import static java.text.MessageFormat.format;
