+++ ladder/jaeeun-na-1/src/test/java/org/ooptraining/game/GameContextTest.java
@@ -1,7 +1,6 @@
 package org.ooptraining.game;
 
 import org.junit.jupiter.api.Test;
-import org.ooptraining.Participant;
 
 import java.util.Arrays;
