+++ ladder/jaeeun-na-1/src/test/java/org/ooptraining/query/QueryProcessorTest.java
@@ -2,7 +2,7 @@ package org.ooptraining.query;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.ooptraining.Participant;
+import org.ooptraining.game.Participant;
 import org.ooptraining.io.StandardInputOutput;
 import org.ooptraining.game.GameContext;
 import org.ooptraining.util.IOUtils;
