+++ ladder/jaeeun-na-1/src/test/java/org/ooptraining/render/RendererTest.java
@@ -2,7 +2,7 @@ package org.ooptraining.render;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.ooptraining.Participant;
+import org.ooptraining.game.Participant;
 import org.ooptraining.game.GameContext;
 
 import java.util.Arrays;
